requirements.txt
# Dev Platform Requirements
requests>=2.31.0
playwright>=1.40.0

# Mastermind Agent System
anthropic>=0.40.0
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
